envoy.base.utils: Add typed util
