+++ envoy.base.utils/VERSION
@@ -1,1 +1,1 @@
-0.0.4-dev
+0.0.4
