+++ envoy.base.utils/setup.cfg
@@ -27,7 +27,9 @@ classifiers =
 python_requires = >=3.5
 py_modules = envoy.base.utils
 packages = find_namespace:
-install_requires = pyyaml
+install_requires =
+    pyyaml
+    trycast
 
 [options.extras_require]
 test =
