+++ mypy-abstracts/VERSION
@@ -1,1 +1,1 @@
-0.0.6
+0.0.7-dev
